pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "delphi"
version = "0.2.0"
description = "Automated Interpretability"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["interpretability", "explainable-ai"]
dependencies = [
    "datasets",
    "orjson",
    "sparsify@git+https://github.com/EleutherAI/sparsify",
    "safetensors",
    "simple_parsing",
    "jaxtyping",
    "fire",
    "blobfile",
    "bitsandbytes",
    "flask",
    "vllm",
    "aiofiles",
    "sentence_transformers",
    "kaleido==0.2.1",
    "plotly==5.0.0rc2",
    "anyio>=4.8.0",
    "triton>=3.1.0",
    "setuptools>=75.8.2",
    "matplotlib>=3.10.1",
    "seaborn>=0.13.2",
    "ipdb>=0.13.13",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.pyright]
include = ["delphi*"]
reportPrivateImportUsage = false

[tool.setuptools.packages.find]
include = ["delphi*"]

[tool.ruff]
# TODO: Clean up or remove experiments folder.
exclude = ["experiments", "**/prompts/*", "**/prompts.py","**/simulator.py"]
# Avoid automatically removing unused imports in __init__.py files.
# Such imports will be flagged with a dedicated message suggesting
# that the import is either added to the module's __all__ symbol
ignore-init-module-imports = true
# Ignore F821 (undefined name) errors, as they are common in type annotations.
ignore = ["F722", "F821"]
# Same as Black.
line-length = 88
# Enable pycodestyle (`E`), Pyflakes (`F`), and isort (`I`) codes
# See https://beta.ruff.rs/docs/rules/ for more possible rules
select = ["E", "F", "I"]
